pyproject.toml
[tool.poetry]
name = "basic_django"
version = "0.1.0"
description = ""
authors = ["usmanamjad048 <1999usmanamjad@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.3"
pip = "^24.0"
install = "^1.3.5"
djangorestframework = "^3.15.1"
pillow = "^10.2.0"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
django-rest-auth = "^0.9.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
